+++ app/src/main/java/dev/leonlatsch/photok/gallery/albums/detail/ui/AlbumDetailViewModel.kt
@@ -23,8 +23,7 @@ import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.leonlatsch.photok.gallery.albums.domain.AlbumRepository
-import dev.leonlatsch.photok.gallery.ui.PhotoTile
-import dev.leonlatsch.photok.model.database.entity.Photo
+import dev.leonlatsch.photok.gallery.ui.components.PhotoTile
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
