+++ app/src/main/java/dev/leonlatsch/photok/gallery/albums/ui/compose/AlbumsPlaceholder.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import dev.leonlatsch.photok.gallery.albums.ui.AlbumsUiEvent
-import dev.leonlatsch.photok.gallery.ui.compose.MagicFab
+import dev.leonlatsch.photok.ui.components.MagicFab
 import dev.leonlatsch.photok.ui.theme.AppTheme
