+++ app/src/main/java/dev/leonlatsch/photok/gallery/ui/compose/GalleryPlaceholder.kt
@@ -39,7 +39,8 @@ import androidx.compose.ui.unit.dp
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.gallery.ui.GalleryUiEvent
 import dev.leonlatsch.photok.ui.theme.AppTheme
-import dev.leonlatsch.photok.uicomponnets.compose.AppName
+import dev.leonlatsch.photok.ui.components.AppName
+import dev.leonlatsch.photok.ui.components.MagicFab
 
 @Composable
 fun GalleryPlaceholder(handleUiEvent: (GalleryUiEvent) -> Unit) {
