+++ app/src/main/java/dev/leonlatsch/photok/ui/components/MagicFab.kt
@@ -14,16 +14,14 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.gallery.ui.compose
+package dev.leonlatsch.photok.ui.components
 
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.ui.theme.AppTheme
